Rename u/u5 package to gddo.
Also improve its API to be nicer and more future-proof.
